Refactor: Move strategy greation into a factory, make funnel count aggregation function parametric
